DBZ-5758 Introduce common modules for common code
