+++ debezium-quarkus-outbox-reactive/integration-tests/src/main/resources/application.properties
@@ -2,8 +2,6 @@ quarkus.datasource.db-kind=postgresql
 quarkus.datasource.username=postgres
 quarkus.datasource.password=postgres
 quarkus.datasource.reactive.url = vertx-reactive:postgresql://localhost:5432/postgres
-#quarkus.datasource.jdbc.url=jdbc:postgresql://localhost:5432/postgres
-#quarkus.hibernate-orm.dialect=org.hibernate.dialect.PostgreSQLDialect
 quarkus.hibernate-orm.database.generation=drop-and-create
 quarkus.hibernate-orm.log.sql=true
 pull.pause.timeout=120
