+++ debezium-quarkus-outbox/deployment-common/src/main/java/io/debezium/outbox/quarkus/deployment/OutboxEventEntityBuildItem.java
