+++ debezium-quarkus-outbox/runtime-common/src/main/java/io/debezium/outbox/quarkus/ExportedEvent.java
