+++ debezium-quarkus-outbox/runtime-common/src/main/java/io/debezium/outbox/quarkus/internal/OutboxConstants.java
@@ -15,6 +15,8 @@ public final class OutboxConstants {
     public static final String OUTBOX_ENTITY_HBMXML = "META-INF/OutboxEvent.hbm.xml";
     public static final String OUTBOX_ENTITY_FULLNAME = "io.debezium.outbox.quarkus.internal.OutboxEvent";
 
+    public static final String TRACING_SPAN_CONTEXT = "tracingspancontext";
+
     private OutboxConstants() {
     }
 }
