+++ debezium-quarkus-outbox/runtime/src/main/java/io/debezium/outbox/quarkus/internal/AbstractEventDispatcher.java
@@ -7,32 +7,19 @@ package io.debezium.outbox.quarkus.internal;
 
 import static io.debezium.outbox.quarkus.internal.OutboxConstants.OUTBOX_ENTITY_FULLNAME;
 
-import java.util.HashMap;
 import java.util.Map;
 
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
 
 import org.hibernate.Session;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import io.debezium.outbox.quarkus.ExportedEvent;
 
 /**
  * Abstract base class for the Debezium Outbox {@link EventDispatcher} contract.
  *
  * @author Chris Cranford
  */
-public abstract class AbstractEventDispatcher implements EventDispatcher {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(AbstractEventDispatcher.class);
-
-    protected static final String TIMESTAMP = "timestamp";
-    protected static final String PAYLOAD = "payload";
-    protected static final String TYPE = "type";
-    protected static final String AGGREGATE_ID = "aggregateId";
-    protected static final String AGGREGATE_TYPE = "aggregateType";
+public abstract class AbstractEventDispatcher extends AbstractEventWriter<Void> implements EventDispatcher {
 
     @Inject
     EntityManager entityManager;
@@ -48,36 +35,20 @@ public abstract class AbstractEventDispatcher implements EventDispatcher {
      *
      * @param dataMap the data map, should never be {@code null}
      */
-    protected void persist(Map<String, Object> dataMap) {
+    protected Void persist(Map<String, Object> dataMap) {
         // Unwrap to Hibernate session and save
         Session session = entityManager.unwrap(Session.class);
         session.save(OUTBOX_ENTITY_FULLNAME, dataMap);
         session.setReadOnly(dataMap, true);
+        remove(dataMap, session);
 
-        // Remove entity if the configuration deems doing so, leaving useful
-        // for debugging
+        return null;
+    }
+
+    private void remove(Map<String, Object> dataMap, Session session) {
+        // Remove entity if the configuration deems doing so, leaving useful for debugging
         if (config.removeAfterInsert) {
             session.delete(OUTBOX_ENTITY_FULLNAME, dataMap);
         }
-    }
-
-    protected Map<String, Object> getDataMapFromEvent(ExportedEvent<?, ?> event) {
-        final HashMap<String, Object> dataMap = new HashMap<>();
-        dataMap.put(AGGREGATE_TYPE, event.getAggregateType());
-        dataMap.put(AGGREGATE_ID, event.getAggregateId());
-        dataMap.put(TYPE, event.getType());
-        dataMap.put(PAYLOAD, event.getPayload());
-        dataMap.put(TIMESTAMP, event.getTimestamp());
-
-        for (Map.Entry<String, Object> additionalFields : event.getAdditionalFieldValues().entrySet()) {
-            if (dataMap.containsKey(additionalFields.getKey())) {
-                LOGGER.error("Outbox entity already contains field with name '{}', additional field mapping skipped",
-                        additionalFields.getKey());
-                continue;
-            }
-            dataMap.put(additionalFields.getKey(), additionalFields.getValue());
-        }
-
-        return dataMap;
     }
 }
