+++ debezium-quarkus-outbox/runtime/src/main/java/io/debezium/outbox/quarkus/internal/DebeziumTracerEventDispatcher.java
@@ -31,7 +31,6 @@ import io.opentracing.tag.Tags;
 @ApplicationScoped
 public class DebeziumTracerEventDispatcher extends AbstractEventDispatcher {
 
-    public static final String TRACING_SPAN_CONTEXT = "tracingspancontext";
     private static final String OPERATION_NAME = "outbox-write";
     private static final String TRACING_COMPONENT = "debezium";
 
@@ -63,7 +62,7 @@ public class DebeziumTracerEventDispatcher extends AbstractEventDispatcher {
 
             // Define the entity map-mode object using property names and values
             final Map<String, Object> dataMap = getDataMapFromEvent(event);
-            dataMap.put(TRACING_SPAN_CONTEXT, exportedSpanData.export());
+            dataMap.put(OutboxConstants.TRACING_SPAN_CONTEXT, exportedSpanData.export());
             persist(dataMap);
         }
         finally {
